+++ package.json
@@ -9,7 +9,8 @@
     "bootstrap": "4.3.1",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
-    "react-scripts": "3.4.0"
+    "react-scripts": "3.4.0",
+    "redux": "4.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
